Layer generation
Finished the layergeneration, it takes a 4×16 table and spits out a
layer. Fairly pessimistic as it checks if the key exist and the size is
good. Using a different matrix requires altering this check and the
error handling in the `pcase` as it assumes a table 16 cells wide.

Also got rid of the big forsaying as it's kinda outdated, need to add
some stuff soon:tm: